src/requirements.txt
# Core
Flask>=3.0.0
Flask-CORS>=4.0.0
gunicorn==21.2.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# PDF Generation
fpdf2==2.7.8  # Lightweight PDF librar

# Environment
python-dotenv>=1.0.0

# Build dependencies
setuptools>=70.0.0
wheel>=0.43.0